+++ CHANGELOG.md
@@ -1,7 +1,26 @@
 # Xill Core, Codec, Collection, Date, File, Math, Stream, String, System plugins - Change Log
 All notable changes to this project will be documented in this file
 
-## [3.3.20] - unreleased
+## [3.3.25] - unreleased
+
+
+## [3.3.24] - 16-06-2016
+### Fix
+- Error when parsing java.sql.Date from jdbc plugin [CTC-1644]
+- Licensing of plugins [CTC-1518]
+
+## [3.3.23] - 09-06-2016
+### Fix
+- Sometimes included libraries do not initialize variables [CTC-1631]
+
+## [3.3.22] - 09-06-2016
+
+## [3.3.21] - 09-06-2016
+
+## [3.3.20] - 07-06-2016
+### Change
+- Split System.exec() output by line again [CTC-1466]
+
 ### Fix
 - PhantomJS gives error after starting robot with library ten times [CTC-1543]
 - `File.getText()` keeps stream open [CTC-1609]